requirements.txt
requests
pyyaml
tabulate
RPi.GPIO
RPLCD
